nuevos cambios

+++ dao/service/CuentasService.php
@@ -17,16 +17,14 @@ if($opcion==1) {
     $user = $datos->usuario;
     $cuentas = CuentasDAO::getCuentasByUser($user);
     foreach ($cuentas as $k => $row) {
-        $response[]=array("id"=>$row->getId(),"nombre"=>$row->getNombre(),
-                "tipo_cuenta"=>$row->getTipo_cuenta(),
+        $response[]=array("id"=>$row->getId(),"nombre"=>$row->getNombre(),                
                 "defecto"=>$row->getDefecto(),"cuantia"=>$row->getCuantia(),
-                "fecha"=>$row->getFecha(),"f_fin"=>$row->getFechaFin(),
-                "tipo_cuenta_id"=>$row->getTipo_cuenta_id());
+                "fecha"=>$row->getFecha(),"f_fin"=>$row->getFechaFin());
     }
 }else if($opcion==2) {
     // operación de alta
     $nombre = $datos->nombre;
-    $tipo = $datos->tipo;
+    //$tipo = $datos->tipo;
     $cuantia = $datos->cuantia;
 
     $timestamp = strtotime($datos->fecha);
@@ -42,7 +40,7 @@ if($opcion==1) {
         $fecha_fin = null;
     }
 
-    $result = CuentasDAO::insertCuenta($nombre, $tipo, $usuario, $fecha, $cuantia, $defecto, $fecha_fin);
+    $result = CuentasDAO::insertCuenta($nombre, $usuario, $fecha, $cuantia, $defecto, $fecha_fin);
 
     if($result==true) {
         $response = array('response' => true, 'mensaje' => 'Cuenta dada de alta correctamente.');
@@ -55,7 +53,7 @@ if($opcion==1) {
 
     $id= $datos->id;
     $nombre = $datos->nombre;
-    $tipo = $datos->tipo;
+    //$tipo = $datos->tipo;
     $cuantia = $datos->cuantia;
 
     $timestamp = strtotime($datos->fecha);
@@ -76,7 +74,7 @@ if($opcion==1) {
         $result = CuentasDAO::updateCuentaDefecto();
     }
 
-    $result = CuentasDAO::updateCuentaById($id, $nombre, $tipo, $fecha, $cuantia, $defecto, $fecha_fin);
+    $result = CuentasDAO::updateCuentaById($id, $nombre, $fecha, $cuantia, $defecto, $fecha_fin);
 
     if($result==true) {
         $response = array('response' => true, 'mensaje' => 'Operación realizada correctamente.');
@@ -84,7 +82,7 @@ if($opcion==1) {
         $response = array('response' => true, 'mensaje' => 'Se ha producido un error en la operación.');
     }
 
-}else {
+}else if($opcion==5){
     // operación de eliminación
     $id = $datos->id;
     $result = CuentasDAO::deleteCuentaById($id);
@@ -94,6 +92,39 @@ if($opcion==1) {
     }else {
         $response = array('response' => true, 'mensaje' => 'Se ha producido un error en la operación.');
     }
+}else{
+    // operación de búsqueda
+    $nombre=null;
+    $cuantiaDesde=null;
+    $cuantiaHasta=null;
+    $fechaDesde=null;
+    $fechaHasta=null;
+
+    $usuario = $datos->usuario;
+
+    if(isset($datos->nombre)){
+        $nombre = $datos->nombre;
+    }
+    if(isset($datos->cuantiaDesde)){
+        $cuantiaDesde = $datos->cuantiaDesde;
+    }
+    if(isset($datos->cuantiaHasta)){
+        $cuantiaHasta = $datos->cuantiaHasta;
+    }
+    if(isset($datos->fechaDesde)){        
+        $fechaDesde = $datos->fechaDesde;
+    }
+    if(isset($datos->fechaHasta)){
+        $fechaHasta = $datos->fechaHasta;
+    }
+
+    $cuentas = CuentasDAO::getCuentasByParams($usuario, $nombre, $cuantiaDesde, $cuantiaHasta, $fechaDesde, $fechaHasta);
+    foreach ($cuentas as $k => $row) {
+        $response[]=array("id"=>$row->getId(),"nombre"=>$row->getNombre(),
+                "defecto"=>$row->getDefecto(),"cuantia"=>$row->getCuantia(),
+                "fecha"=>$row->getFecha(),"f_fin"=>$row->getFechaFin());
+    }
+
 }
 
 //no hacer return, sino echo
